Check all bundles with PHPStan and clean up the inline phpDocs (see #6271)
Description
-----------



Commits
-------

5377fb82 Check all bundles with PHPStan
a64fa81c Remove inline phpDocs that are no longer necessary
293c58f6 Fix the PageUrlListenerTest

+++ test-case/src/ContaoDatabaseTrait.php
@@ -40,7 +40,7 @@ trait ContaoDatabaseTrait
                 $params = [
                     'driver' => 'pdo_mysql',
                     'host' => getenv('DB_HOST'),
-                    'port' => getenv('DB_PORT'),
+                    'port' => (int) getenv('DB_PORT'),
                     'user' => getenv('DB_USER'),
                     'password' => getenv('DB_PASS'),
                     'dbname' => getenv('DB_NAME'),
